+++ spec/csvy_spec.rb
@@ -25,8 +25,16 @@ describe Jekyll::Csvy do
       end
     end
 
+    describe "convert csv with multiple lines in a block" do
+      csv = "Header,Options\nOne,\"* 1\n* 2\n* 3\"\n"
+      html = "<table style=\"width:32%;\">\n<colgroup>\n<col width=\"12%\" />\n<col width=\"19%\" />\n</colgroup>\n<thead>\n<tr class=\"header\">\n<th>Header</th>\n<th>Options</th>\n</tr>\n</thead>\n<tbody>\n<tr class=\"odd\">\n<td><p>One</p></td>\n<td><ul>\n<li>1</li>\n<li>2</li>\n<li>3</li>\n</ul></td>\n</tr>\n</tbody>\n</table>\n"
+      it "should convert csv" do
+        expect(subject.convert(csv)).to eq(html)
+      end
+    end
+
     describe "convert invalid csv" do
-      csv = "This is a heading\n1,2,3\n"
+      csv = "Header\n1,2,3\n"
       it "should raise error" do
         expect { subject.convert(csv) }.to raise_error(Jekyll::Errors::FatalException, /Conversion failed with error/)
       end