+++ CHANGELOG.md
@@ -1,1 +1,11 @@
+## v.0.4 (May 19, 2016)
 
+[jekyll-csvy 0.4](https://github.com/datacite/jekyll-csvy/releases/tag/v.0.4) was released on May 19, 2016 with the following changes:
+
+* support for block elements (lists, etc.) in table cells
+
+## v.0.3 (May 18, 2016)
+
+[jekyll-csvy 0.3](https://github.com/datacite/jekyll-csvy/releases/tag/v.0.3) was released on May 18, 2016:
+
+* initial release